add teardown hook to tests for better stability update fixture to have multiple qdc records and filenames add local python2.7 support files to gitignore add note on handling filename on same line as qdc tag
