+++ tests/fixtures/qdc1.xml
@@ -1,3 +1,4 @@
+FILENAME test1
 <qualifieddc xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:marcrel="http://www.loc.gov/marc.relators/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.loc.gov/marc.relators/ http://imlsdcc2.grainger.illinois.edu/registry/marcrel.xsd" xsi:noNamespaceSchemaLocation="http://dublincore.org/schemas/xmls/qdc/2008/02/11/qualifieddc.xsd">
   <dc:title>Carved flint macehead, Eastern tomb chamber, Knowth </dc:title>
   <dc:description>Maesmawr-type ovoid flint macehead, with cylindrical perforation towards narrow end. All six surfaces are decorated. Discovered in the right-hand recess of the chamber of Knowth Tomb 1B East.</dc:description>
@@ -15,4 +16,9 @@
   <dc:source>Photograph</dc:source>
   <dc:format>Jpeg</dc:format>
   <dc:rights>Copyright National Monuments Service, Department of Culture, Heritage and the Gaeltacht.</dc:rights>
+</qualifieddc>
+
+FILENAME test2
+<qualifieddc xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:marcrel="http://www.loc.gov/marc.relators/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.loc.gov/marc.relators/ http://imlsdcc2.grainger.illinois.edu/registry/marcrel.xsd" xsi:noNamespaceSchemaLocation="http://dublincore.org/schemas/xmls/qdc/2008/02/11/qualifieddc.xsd">
+  <dc:title>Foo</dc:title>
 </qualifieddc>
